Moved all files to packages to look cleaner
